functest-prepare: auto-generate model name when not provided
Fixes #137
